round Irms to 2 decimals

+++ include/webthing.h
@@ -3,4 +3,4 @@
 #include <Thing.h>
 
 void setupWebThing(String deviceName);
-void updateThingProperties(double current, double voltage, double power);
+void updateThingProperties(float current, float voltage, float power);
